Move useConfigFiles method to Container module
