+++ src/Module/Nette.php
@@ -11,11 +11,10 @@
 namespace Arachne\Codeception\Module;
 
 use Arachne\Codeception\Connector\Nette as NetteConnector;
-use Arachne\Codeception\Module\Container as ContainerModule;
 use Codeception\Lib\Framework;
 use Codeception\TestInterface;
-use Nette\DI\Container;
-use Nette\DI\MissingServiceException;
+use Nette\Http\IRequest;
+use Nette\Http\IResponse;
 
 /**
  * @author Jáchym Toušek <enumag@gmail.com>
@@ -25,21 +24,6 @@ class Nette extends Framework
     protected $config = [
         'followRedirects' => true,
     ];
-
-    /**
-     * @var array
-     */
-    private $configFiles;
-
-    /**
-     * @var Container
-     */
-    private $container;
-
-    /**
-     * @var callable
-     */
-    private $containerAccessor;
 
     /**
      * @var string
@@ -54,28 +38,13 @@ class Nette extends Framework
     public function _before(TestInterface $test)
     {
         $this->configFiles = null;
-        $this->container = null;
-        $this->containerAccessor = function () {
-            if (!$this->container) {
-                $this->container = $this->getModule(ContainerModule::class)->createContainer();
-            }
-
-            return $this->container;
-        };
-
         $this->client = new NetteConnector();
-        $this->client->setContainerAccessor($this->containerAccessor);
+        $this->client->setContainerAccessor(function () {
+            return $this->getModule(Container::class)->getContainer();
+        });
         $this->client->followRedirects($this->config['followRedirects']);
 
         parent::_before($test);
-    }
-
-    public function useConfigFiles(array $configFiles)
-    {
-        if ($this->container) {
-            $this->fail('Can\'t set configFiles after the container is created.');
-        }
-        $this->configFiles = $configFiles;
     }
 
     public function _after(TestInterface $test)
@@ -94,9 +63,8 @@ class Nette extends Framework
         if ($this->config['followRedirects']) {
             $this->fail('Method seeRedirectTo only works when followRedirects option is disabled');
         }
-        $container = call_user_func($this->containerAccessor);
-        $request = $container->getByType('Nette\Http\IRequest');
-        $response = $container->getByType('Nette\Http\IResponse');
+        $request = $this->getModule(Container::class)->grabService(IRequest::class);
+        $response = $this->getModule(Container::class)->grabService(IResponse::class);
         if ($response->getHeader('Location') !== $request->getUrl()->getHostUrl().$url && $response->getHeader('Location') !== $url) {
             $this->fail('Couldn\'t confirm redirect target to be "'.$url.'", Location header contains "'.$response->getHeader('Location').'".');
         }
